add:投稿詳細ページにいいねボタンを追加
